split draw-state file into store and manager
this helps avoid creating an import cycle when adding to the unload
handler in the next commit.
